Saved recipe: TEST.compute_nyc_trips_tip_percentage

+++ recipes/compute_nyc_trips_tip_percentage.sql
@@ -2,7 +2,7 @@ SELECT
     *,
     CASE 
         WHEN total_amount = 0 THEN 0
-        ELSE (tip_amount / total_amount) * 100
+        ELSE (tip_amount::float / total_amount::float) * 100
     END AS tip_percentage
 FROM 
     "public"."SQL_nyc_trips_copy"
